refactor: move notation under configuration/
